feat: clean up tests
